Refactor wind system to use windPower int parameter!

+++ ProjectWinter/SourceCode/project_winter/src/snowfall.h
@@ -17,7 +17,7 @@ public:
     ~Snowfall();
 
     // Call every frame (fit the camera's view box!)
-    void update(float deltaTime, const mat4& view, const mat4& proj, float windPower);
+    void update(float deltaTime, const mat4& view, const mat4& proj, int windPower);
 
     // Call every frame!
     void draw(const mat4& view, const mat4& proj);
